Refactoring Validation

+++ src/Product/Deleteproduct.js
@@ -2,6 +2,7 @@ import React from 'react';
 import './Addproduct.css';
 import { Container, Col, Form, FormGroup, Label } from 'reactstrap';
 import axios from 'axios';
+let Isvalid;
 
 export const Productlist = (props) => (
     <div className="form-group">
@@ -10,7 +11,7 @@ export const Productlist = (props) => (
             className="form-control"
             name="{props.name}"
             onChange={props.onChange} style={{ float: 'left', paddingRight: '5px', width: '30%' }} >
-            <option defaultValue>Select {props.name}</option>
+            <option defaultValue>Select Product </option>
             {props.options.map((item, index) => (
                 <option key={index} value={item.id}>
                     {item.name}
@@ -39,15 +40,27 @@ export default class FetchProductDetails extends React.Component {
     }
 
     DeleteProduct = () => {
-        axios.delete(`https://e-auction-api-gate-way.azurewebsites.net/apigateway/e-auction/api/v1/seller/delete/${this.state.product}`)
-            .then(json => {
-                if (json.status = 200) {
-                    alert("Product Delete Successfully");
-                    window.location.href = 'https://e-auction-web-app.azurewebsites.net/';
-                }
-            }).catch(err => {
-                alert(err.response.data);
-            })
+
+        if (this.state.product === '') {
+            alert("Please choose product");
+            Isvalid = false;
+        }
+        else {
+            Isvalid = true;
+        }
+
+        if (Isvalid) {
+
+            axios.delete(`https://e-auction-api-gate-way.azurewebsites.net/apigateway/e-auction/api/v1/seller/delete/${this.state.product}`)
+                .then(json => {
+                    if (json.status = 200) {
+                        alert("Product Delete Successfully");
+                        window.location.href = 'https://e-auction-web-app.azurewebsites.net/';
+                    }
+                }).catch(err => {
+                    alert(err.response.data);
+                })
+        }
     }
 
     onChange = (event) => {
